fixed type issues
