+++ apps/api/src/modules/v1/videos/model.ts
@@ -1,65 +1,92 @@
-import { z } from "zod";
+import { createSelectSchema } from "drizzle-typebox";
+import { t } from "elysia";
 
-export const VideoModel = {
-  // Common Params (Client sends ID in URL)
-  params: z.object({
-    id: z.string(),
-  }),
+import { videos } from "@vidcastx/database/schema/video-schema";
 
-  // Pagination Query
-  pagination: z.object({
-    page: z.coerce.number().min(1).default(1).optional(),
-    limit: z.coerce.number().min(1).max(100).default(10).optional(),
-  }),
+/**
+ * Schemas derived from Drizzle table using drizzle-typebox
+ * These are the source of truth for validation and OpenAPI docs
+ */
 
-  // Create Draft
-  create: z.object({
-    title: z.string().min(3).max(100).optional(),
-    folderId: z.string().optional(),
-    filename: z.string().min(1),
-    contentType: z.string().regex(/^video\//, "Must be a video file"),
-    description: z.string().optional(),
-  }),
+// Full video response (from database)
+const _videoSelect = createSelectSchema(videos);
+export const VideoResponse = t.Omit(_videoSelect, ["masterAccessUrl", "metadata"]);
 
-  // Update Metadata
-  update: z.object({
-    title: z.string().min(3).max(100).optional(),
-    description: z.string().optional(),
-    visibility: z.enum(["public", "private", "unlisted"]).optional(),
-    schedule: z.iso.datetime().optional(), // ISO Date string
-  }),
+// Create draft video — filename is API-only (used to derive S3 key/extension), not stored in DB
+export const CreateVideoBody = t.Object({
+  filename: t.String({ minLength: 1 }),
+  title: t.Optional(t.String({ minLength: 3, maxLength: 200 })),
+  folderId: t.Optional(t.Nullable(t.String())),
+  description: t.Optional(t.String()),
+});
 
-  // Multipart
-  multipartInit: z.object({
-    contentType: z.string(),
-  }),
+// Update video metadata
+export const UpdateVideoBody = t.Partial(t.Pick(_videoSelect, ["title", "description", "visibility", "scheduledAt"]));
 
-  multipartSign: z.object({
-    uploadId: z.string(),
-    partNumber: z.coerce.number(),
-  }),
+// Pagination query
+export const PaginationQuery = t.Object({
+  page: t.Optional(t.Numeric({ minimum: 1, default: 1 })),
+  limit: t.Optional(t.Numeric({ minimum: 1, maximum: 100, default: 10 })),
+});
 
-  multipartComplete: z.object({
-    uploadId: z.string(),
-    parts: z.array(
-      z.object({
-        ETag: z.string(),
-        PartNumber: z.number(),
-      }),
-    ),
-  }),
+// Video ID param
+export const VideoIdParam = t.Object({
+  id: t.String({ pattern: "^vid_" }),
+});
 
-  multipartListParts: z.object({
-    uploadId: z.string(),
-  }),
+// Multipart upload init
+export const MultipartInitBody = t.Object({
+  contentType: t.String({ pattern: "^video/" }),
+});
 
-  multipartAbort: z.object({
-    uploadId: z.string(),
-  }),
+// Multipart sign part
+export const MultipartSignQuery = t.Object({
+  uploadId: t.String(),
+  partNumber: t.Numeric({ minimum: 1 }),
+});
 
-  updateProcessingStatus: z.object({
-    status: z.enum({ processing: "processing", ready: "ready", failed: "failed" }),
-    playbackUrl: z.string().optional(),
-    errorReason: z.string().optional(),
-  }),
-};
+// Multipart complete
+export const MultipartCompleteBody = t.Object({
+  uploadId: t.String(),
+  parts: t.Array(
+    t.Object({
+      ETag: t.String(),
+      PartNumber: t.Numeric(),
+    }),
+  ),
+});
+
+// Multipart list parts
+export const MultipartListPartsQuery = t.Object({
+  uploadId: t.String(),
+});
+
+// Multipart abort
+export const MultipartAbortBody = t.Object({
+  uploadId: t.String(),
+});
+
+// List response
+export const VideoListResponse = t.Object({
+  page: t.Numeric(),
+  limit: t.Numeric(),
+  total: t.Numeric(),
+  videos: t.Array(VideoResponse),
+});
+
+// Create response
+export const CreateVideoResponse = t.Object({
+  status: t.Literal("created"),
+  data: VideoResponse,
+});
+
+// Error response
+export const ErrorResponse = t.Object({
+  error: t.String(),
+});
+
+// Generic success response
+export const SuccessResponse = t.Object({
+  status: t.String(),
+  videoId: t.String(),
+});
