+++ apps/dashboard/package.json
@@ -15,10 +15,10 @@
   },
   "dependencies": {
     "@elysiajs/eden": "^1.4.8",
-    "@t3-oss/env-core": "^0.13.10",
     "@hookform/resolvers": "^5.2.2",
     "@hugeicons/core-free-icons": "^3.3.0",
     "@hugeicons/react": "^1.1.5",
+    "@t3-oss/env-core": "^0.13.10",
     "@tanstack/react-form": "^1.28.4",
     "@tanstack/react-query": "^5.90.21",
     "@tanstack/react-query-devtools": "^5.91.3",
